refactor: split zigzag into multiple files in its own module
