+++ storage-proofs/src/zigzag/mod.rs
@@ -1,0 +1,14 @@
+#[macro_use]
+mod macros;
+
+mod challenges;
+mod column;
+mod column_proof;
+mod encoding_proof;
+mod graph;
+mod hash;
+mod proof;
+
+pub use self::challenges::{ChallengeRequirements, LayerChallenges};
+pub use self::graph::{ZigZagBucketGraph, ZigZagGraph, EXP_DEGREE};
+pub use self::proof::ZigZagDrgPoRep;
